refactor: Extract persistance functions from client.js
